Fix duplicate zip entry error
